# tools/modules/testing: Make testing run in parallel