--- a/vlib/benchmark/benchmark.v
+++ b/vlib/benchmark/benchmark.v
@@ -2,6 +2,7 @@ module benchmark
 
 import time
 import term
+
 /*
 Example usage of this module:
 ```
@@ -25,110 +26,124 @@ println( bmark.total_message('remarks about the benchmark') )
 ```
 */
 
-
 const (
 	BOK = term.ok_message('OK')
 	BFAIL = term.fail_message('FAIL')
 )
 
-pub struct Benchmark {
+pub struct ConcurrentBenchmark {
 pub mut:
 	bench_start_time i64
 	bench_end_time   i64
-	step_start_time  i64
-	step_end_time    i64
 	ntotal           int
 	nok              int
 	nfail            int
 	verbose          bool
 	nexpected_steps  int
-	cstep            int
 	bok              string
 	bfail            string
 }
 
-pub fn new_benchmark() Benchmark {
-	return Benchmark{
+pub struct ConcurrentInstance {
+pub mut:
+	benchmark       &ConcurrentBenchmark
+	step_start_time i64
+	step_end_time   i64
+	step_number     int
+}
+
+pub fn new_concurrent_benchmark() ConcurrentBenchmark {
+	return ConcurrentBenchmark{
 		bench_start_time: benchmark.now()
 		verbose: true
 	}
 }
 
-pub fn (b mut Benchmark) set_total_expected_steps(n int) {
+pub fn (b mut ConcurrentBenchmark) set_total_expected_steps(n int) {
 	b.nexpected_steps = n
 }
 
-pub fn (b mut Benchmark) stop() {
+pub fn (b mut ConcurrentBenchmark) stop() {
 	b.bench_end_time = benchmark.now()
 }
 
-pub fn (b mut Benchmark) step() {
-	b.step_start_time = benchmark.now()
-	b.cstep++
+pub fn (b &ConcurrentBenchmark) step() ConcurrentInstance {
+	return ConcurrentInstance {
+		benchmark: b
+		step_start_time: benchmark.now()
+	}
 }
 
-pub fn (b mut Benchmark) fail() {
+pub fn (b mut ConcurrentInstance) fail() {
 	b.step_end_time = benchmark.now()
-	b.ntotal++
-	b.nfail++
+	b.benchmark.ntotal++
+	b.benchmark.nfail++
+	b.step_number = b.benchmark.ntotal
 }
 
-pub fn (b mut Benchmark) ok() {
+pub fn (b mut ConcurrentInstance) ok() {
 	b.step_end_time = benchmark.now()
-	b.ntotal++
-	b.nok++
+	b.benchmark.ntotal++
+	b.benchmark.nok++
+	b.step_number = b.benchmark.ntotal
 }
 
-pub fn (b mut Benchmark) fail_many(n int) {
+pub fn (b mut ConcurrentInstance) fail_many(n int) {
 	b.step_end_time = benchmark.now()
-	b.ntotal += n
-	b.nfail += n
+	b.benchmark.ntotal += n
+	b.benchmark.nfail += n
+	b.step_number = b.benchmark.ntotal
 }
 
-pub fn (b mut Benchmark) ok_many(n int) {
+pub fn (b mut ConcurrentInstance) ok_many(n int) {
 	b.step_end_time = benchmark.now()
-	b.ntotal += n
-	b.nok += n
+	b.benchmark.ntotal += n
+	b.benchmark.nok += n
+	b.step_number = b.benchmark.ntotal
 }
 
-pub fn (b mut Benchmark) neither_fail_nor_ok() {
+pub fn (b mut ConcurrentInstance) neither_fail_nor_ok() {
 	b.step_end_time = benchmark.now()
+	b.benchmark.ntotal++
+	b.step_number = b.benchmark.ntotal
 }
 
-pub fn (b &Benchmark) step_message_with_label(label string, msg string) string {
+pub fn (b &ConcurrentInstance) step_message_with_label(label string, msg string) string {
 	mut timed_line := ''
-	if b.nexpected_steps > 0 {
+	expected_steps := b.benchmark.nexpected_steps
+
+	if expected_steps > 0 {
 		mut sprogress := ''
-		if b.nexpected_steps < 10 {
-			sprogress = '${b.cstep:1d}/${b.nexpected_steps:1d}'
+		if expected_steps < 10 {
+			sprogress = '${b.step_number:1d}/${expected_steps:1d}'
 		}
-		if b.nexpected_steps >= 10 && b.nexpected_steps < 100 {
-			sprogress = '${b.cstep:2d}/${b.nexpected_steps:2d}'
+		if expected_steps >= 10 && expected_steps < 100 {
+			sprogress = '${b.step_number:2d}/${expected_steps:2d}'
 		}
-		if b.nexpected_steps >= 100 && b.nexpected_steps < 1000 {
-			sprogress = '${b.cstep:3d}/${b.nexpected_steps:3d}'
+		if expected_steps >= 100 && expected_steps < 1000 {
+			sprogress = '${b.step_number:3d}/${expected_steps:3d}'
 		}
-		timed_line = b.tdiff_in_ms('[${sprogress}] $msg', b.step_start_time, b.step_end_time)
+		timed_line = b.benchmark.tdiff_in_ms('[${sprogress}] $msg', b.step_start_time, b.step_end_time)
 	}
 	else {
-		timed_line = b.tdiff_in_ms(msg, b.step_start_time, b.step_end_time)
+		timed_line = b.benchmark.tdiff_in_ms(msg, b.step_start_time, b.step_end_time)
 	}
 	return '${label:-5s}${timed_line}'
 }
 
-pub fn (b &Benchmark) step_message(msg string) string {
+pub fn (b &ConcurrentInstance) step_message(msg string) string {
 	return b.step_message_with_label('', msg)
 }
 
-pub fn (b &Benchmark) step_message_ok(msg string) string {
+pub fn (b &ConcurrentInstance) step_message_ok(msg string) string {
 	return b.step_message_with_label(BOK, msg)
 }
 
-pub fn (b &Benchmark) step_message_fail(msg string) string {
+pub fn (b &ConcurrentInstance) step_message_fail(msg string) string {
 	return b.step_message_with_label(BFAIL, msg)
 }
 
-pub fn (b &Benchmark) total_message(msg string) string {
+pub fn (b &ConcurrentBenchmark) total_message(msg string) string {
 	mut tmsg := '${msg}\n                 ok, fail, total = ' + term.ok_message('${b.nok:5d}') + ', ' + if b.nfail > 0 { term.fail_message('${b.nfail:5d}') } else { '${b.nfail:5d}' } + ', ' + '${b.ntotal:5d}'
 	if b.verbose {
 		tmsg = '<=== total time spent $tmsg'
@@ -136,12 +151,12 @@ pub fn (b &Benchmark) total_message(msg string) string {
 	return '  ' + b.tdiff_in_ms(tmsg, b.bench_start_time, b.bench_end_time)
 }
 
-pub fn (b &Benchmark) total_duration() i64 {
+pub fn (b &ConcurrentBenchmark) total_duration() i64 {
 	return (b.bench_end_time - b.bench_start_time)
 }
 
 // //////////////////////////////////////////////////////////////////
-fn (b &Benchmark) tdiff_in_ms(s string, sticks i64, eticks i64) string {
+fn (b &ConcurrentBenchmark) tdiff_in_ms(s string, sticks i64, eticks i64) string {
 	if b.verbose {
 		tdiff := (eticks - sticks)
 		return '${tdiff:6lld} ms $s'