This PR does two things:
- Make a concurrent friendly version of `vlib/benchmark`
- Make testing (ie. vtest-compiler) run with multiple cores

This allows testing to be completed faster before V compiling is properly multi-threaded.

Before (on commit `b1602c72`):
```
$ time (make && v test-compiler)
   20542 ms <=== total time spent building tools
  130889 ms <=== total time spent running V tests
   29244 ms <=== total time spent building examples
    7874 ms <=== total time spent building examples/hot_reload
Installing a v module took: 7081ms
( make && v test-compiler; )  180.77s user 14.91s system 94% cpu 3:26.69 total
```

After (my CPU has 4 cores):
```
$ time (make && v test-compiler)
    6692 ms <=== total time spent building tools
   59377 ms <=== total time spent running V tests
    8225 ms <=== total time spent building examples
    3502 ms <=== total time spent building examples/hot_reload
Installing a v module took: 6664ms
( make && v test-compiler; )  304.95s user 20.79s system 345% cpu 1:34.17 total
```

While individual test time is longer, it doesn't matter as 8 threads are working on it concurrently.
What I've noticed is that `repl_test.v` was the bottleneck in both cases, using up 42347ms (before) and 48910ms (after, causing only one CPU to be used as other test finished).
Therefore, if we wanted to make tests more parallel, `repl_test.v` is the target to optimize.