@spytheman 2020-01-19
We should aim to minimize the number of .repl tests.
They are convenient, they way they run, they do a recompilation for each *line* of each .repl file.
Only tests that should produce errors should be done in this mode.
All the rest should be moved to ordinary _test.v files.