@spytheman 2020-01-19
Why is there a new file abenchmark_single.v containing Benchmark, while in the old benchmark.v there is a ConcurrentBenchmark ? 

Would not it be clearer if the old one is left in file benchmark.v , while the concurrent one is in a new file?