@spytheman 2020-01-19
Keep in mind that Benchmark is used in several places, including in generated C code by the compiler, and several test suites can not be run concurrently within the _test.v file.
For example os_test.v depends on the order of the tests, so that it cleans up created files.

I would prefer, if the old Benchmark structure remained as it is, if possible.