@medvednikov 2020-01-20
Thanks.

This has been implemented by @spytheman in a different PR.